refactor: added forward ref comp.
