+++ src/components/DetailContainer/DetailContainer.tsx
@@ -1,7 +1,7 @@
-import React, { useState, ChangeEvent } from 'react';
+import React, { useState, ChangeEvent, useContext } from 'react';
 import Box from '@material-ui/core/Box';
 
-import { DetailContextConsumer } from '../../pages/Version';
+import { DetailContext } from '../../pages/Version';
 
 import DetailContainerTabs from './DetailContainerTabs';
 import DetailContainerContent from './DetailContainerContent';
@@ -9,6 +9,8 @@ import { TabPosition } from './tabs';
 
 const DetailContainer: React.FC = () => {
   const [tabPosition, setTabPosition] = useState(TabPosition.README);
+  const detailContext = useContext(DetailContext);
+  const { readMe } = detailContext;
 
   const handleChangeTabPosition = (event: ChangeEvent<{}>) => {
     const eventTarget = event.target as HTMLSpanElement;
@@ -17,14 +19,10 @@ const DetailContainer: React.FC = () => {
   };
 
   return (
-    <DetailContextConsumer>
-      {({ readMe }) => (
-        <Box component="div" display="flex" flexDirection="column" padding={2}>
-          <DetailContainerTabs onChangeTabPosition={handleChangeTabPosition} tabPosition={tabPosition} />
-          <DetailContainerContent readDescription={readMe} tabPosition={tabPosition} />
-        </Box>
-      )}
-    </DetailContextConsumer>
+    <Box component="div" display="flex" flexDirection="column" padding={2}>
+      <DetailContainerTabs onChangeTabPosition={handleChangeTabPosition} tabPosition={tabPosition} />
+      <DetailContainerContent readDescription={readMe} tabPosition={tabPosition} />
+    </Box>
   );
 };
