+++ src/components/Package/Package.tsx
@@ -4,12 +4,12 @@ import BugReport from '@material-ui/icons/BugReport';
 import Grid from '@material-ui/core/Grid';
 import HomeIcon from '@material-ui/icons/Home';
 import ListItem from '@material-ui/core/ListItem';
-import Tooltip from '@material-ui/core/Tooltip';
 
 import { PackageMetaInterface } from 'types/packageMeta';
 import Tag from '../Tag';
 import fileSizeSI from '../../utils/file-size';
 import { formatDate, formatDateDistance } from '../../utils/package';
+import Tooltip from '../../muiComponents/Tooltip';
 import {
   Author,
   Avatar,
